Move `CultureFeedAddressFactory` to `CultureFeed` folder
